requirements.txt
python-dotenv
selenium
webdriver-manager
pygsheets
argparse
ndjson
requests